Implement player argument parsing
This commit adds the utility class Player, which can be constructed from a Minecraft UUID or (if the people package is present) a Wurstmineberg ID. Endpoints expecting player parameters now use this class for argument parsing, which means that clients may now always refer to a player using either ID.

During construction of a Player object, if the API encounters an entry in the people database which has Minecraft usernames set but no Minecraft UUID, it will attempt to retrieve the UUID from the Mojang API and on success write it back to the people database.

Additionally, endpoints returning people.json or parts thereof will now replace the Gravatar email address with a URL of the avatar instead of removing the field altogether.