Remove stringID field from item info endpoints
This field was originally added to the item info objects so that the string ID would always be available, even when the endpoint was called using a numeric ID argument. Since support for numeric IDs as arguments has now been dropped, the field is also dropped for full compatibility with the items.json format.