Make world hostname configurable
In case the API runs on a different hostname from the Minecraft worlds (as is the case on api.dev.wurstmineberg.de), the configuration file now has a field worldHost which will be used to get the list of online players. It defaults to the value of the host field.